Add _meta key for special data
In the future, the _meta key will store all data that
the boardfarm client doesn't need to know, but that the server
needs to know.

For now, we're in a transition of moving some keys there, even
though it's redundant.

This will allow us to save some data even when uploading a new
boardfarm config file.

Signed-off-by: Mike Anderson <mbanderson@uwalumni.com>